Fix adding not-found route
The previous implementation used a simple find-and-replace like solution, which was easy to break when the default file contents changed a bit. This was the case now, as the default blueprint writes the call in a single line, which did not match anymore:

```
Router.map(function () {});
```

Now this uses the standard AST-based ember library for router generation.